- move all payment services to one file - style status page
