+++ src/app/pages/payment/international-payment-status/international-payment-status.component.html
@@ -1,8 +1,15 @@
 <div class="container">
-  <div class="card mb-3">
-    <img class="card-img-top" [src]="imgSrc" alt="Status" />
-    <div class="card-body">
-      <h2 class="card-title">{{ message }}</h2>
+  <div class="card mb-3" style="min-width:220px;width:50%;max-height:360px; margin-top:20px;margin-left:auto; margin-right:auto; text-align:centre;">
+    <div class="row no-gutters">
+      <div class="col-md-4">
+        <img [src]="imgSrc" class="card-img" alt="Status" />
+      </div>
+      <div class="col-md-8">
+        <div class="card-body">
+          <h5 class="card-title">Transaction Status</h5>
+          <p class="card-text text-{{status}}">{{ message }}</p>
+        </div>
+      </div>
     </div>
   </div>
 </div>
