+++ src/app/pages/payment/international-payment-status/international-payment-status.component.scss
@@ -1,3 +1,15 @@
-.text-failure{
-    color:red;
+.text-failure {
+  color: #800000;
+}
+.text-success {
+  color: #006600;
+}
+.card {
+  min-width: 220px;
+  width: 50%;
+  max-height: 360px;
+  margin-top: 20px;
+  margin-left: auto;
+  margin-right: auto;
+  text-align: centre;
 }
