+++ src/app/pages/payment/international-payment/international-payment.component.scss
@@ -1,13 +1,20 @@
+.payment{
+  width: 100%;
+  background-image: url('src/assets/img/HouseBG.jpg');
+  padding-top:20px;
+  padding-bottom:20px;
+}
 .payment-form {
+  background:white;
   width: 65%;
   min-width: 300px;
   margin-left: auto;
   margin-right: auto;
   margin-top: 30px;
   box-shadow: 0 0 5px black;
-}
-.payment-form > .container {
-  margin-bottom: 1em;
+  > .container {
+    margin-bottom: 1em;
+  }
 }
 .heading {
   background: #ffa500;
@@ -43,15 +50,4 @@
   min-width: 280px;
   text-align: center;
   border: none;
-}
-.col {
-  float: none;
-  width: 100%;
-}
-
-@media only screen and (max-width: 600px) {
-  .col {
-    float: none;
-    width: 100%;
-  }
 }
