+++ src/environments/environment.prod.ts
@@ -2,6 +2,7 @@ export const environment = {
   production: true,
   profileUrl:
     'https://landville-backend-web-api.herokuapp.com/api/v1/auth/profile/',
-  api_url: 'https://landville-backend-web-api.herokuapp.com/api/v1',
+  // api_url: 'https://landville-backend-web-api.herokuapp.com/api/v1',
+  api_url: 'http://localhost:4200/api/v1',
 };
