+++ src/environments/environment.ts
@@ -6,5 +6,6 @@ export const environment = {
   production: false,
   profileUrl:
     'https://landville-backend-web-api.herokuapp.com/api/v1/auth/profile/',
-  api_url: 'https://landville-backend-web-api.herokuapp.com/api/v1',
+  // api_url: 'https://landville-backend-web-api.herokuapp.com/api/v1',
+  api_url: 'http://localhost:8000/api/v1',
 };
